Revert "Adds new homepage"
