+++ css/webflow.css
@@ -813,6 +813,7 @@ textarea.w-select {
 .w-widget-twitter-count-shim.w--large {
   width: 36px;
   height: 28px;
+  margin-left: 7px;
 }
 .w-widget-twitter-count-shim.w--large .w-widget-twitter-count-inner {
   font-size: 18px;
@@ -1100,9 +1101,41 @@ textarea.w-select {
 .w-dropdown-link.w--current {
   color: #0082f3;
 }
+.w-nav[data-collapse="all"] .w-dropdown,
+.w-nav[data-collapse="all"] .w-dropdown-toggle {
+  display: block;
+}
+.w-nav[data-collapse="all"] .w-dropdown-list {
+  position: static;
+}
+@media screen and (max-width: 991px) {
+  .w-nav[data-collapse="medium"] .w-dropdown,
+  .w-nav[data-collapse="medium"] .w-dropdown-toggle {
+    display: block;
+  }
+  .w-nav[data-collapse="medium"] .w-dropdown-list {
+    position: static;
+  }
+}
 @media screen and (max-width: 767px) {
+  .w-nav[data-collapse="small"] .w-dropdown,
+  .w-nav[data-collapse="small"] .w-dropdown-toggle {
+    display: block;
+  }
+  .w-nav[data-collapse="small"] .w-dropdown-list {
+    position: static;
+  }
   .w-nav-brand {
     padding-left: 10px;
+  }
+}
+@media screen and (max-width: 479px) {
+  .w-nav[data-collapse="tiny"] .w-dropdown,
+  .w-nav[data-collapse="tiny"] .w-dropdown-toggle {
+    display: block;
+  }
+  .w-nav[data-collapse="tiny"] .w-dropdown-list {
+    position: static;
   }
 }
 /**
@@ -1596,7 +1629,7 @@ textarea.w-select {
   position: relative;
   float: right;
 }
-[data-nav-menu-open] {
+.w--nav-menu-open {
   display: block !important;
   position: absolute;
   top: 100%;
@@ -1620,14 +1653,14 @@ textarea.w-select {
   right: 0;
   width: 100%;
 }
-.w-nav-overlay [data-nav-menu-open] {
+.w-nav-overlay .w--nav-menu-open {
   top: 0;
 }
 .w-nav[data-animation="over-left"] .w-nav-overlay {
   width: auto;
 }
 .w-nav[data-animation="over-left"] .w-nav-overlay,
-.w-nav[data-animation="over-left"] [data-nav-menu-open] {
+.w-nav[data-animation="over-left"] .w--nav-menu-open {
   right: auto;
   z-index: 1;
   top: 0;
@@ -1636,7 +1669,7 @@ textarea.w-select {
   width: auto;
 }
 .w-nav[data-animation="over-right"] .w-nav-overlay,
-.w-nav[data-animation="over-right"] [data-nav-menu-open] {
+.w-nav[data-animation="over-right"] .w--nav-menu-open {
   left: auto;
   z-index: 1;
   top: 0;
@@ -1664,15 +1697,6 @@ textarea.w-select {
 }
 .w-nav[data-collapse="all"] .w-nav-button {
   display: block;
-}
-.w--nav-dropdown-open {
-  display: block;
-}
-.w--nav-dropdown-toggle-open {
-  display: block;
-}
-.w--nav-dropdown-list-open {
-  position: static;
 }
 @media screen and (max-width: 991px) {
   .w-nav[data-collapse="medium"] .w-nav-menu {
@@ -1732,9 +1756,6 @@ textarea.w-select {
 }
 .w-tab-link.w--current {
   background-color: #C8C8C8;
-}
-.w-tab-link:focus {
-  outline: 0;
 }
 .w-tab-content {
   position: relative;
